This is the PHP code for the website!
The website is and the database are connected, the user can register and his data and password (hashed) are stored in the database.

The user can also login and logout at any time and a welcome message with his/hers name will be shown in the header of the page. If the user is not logged in no data from the search bar will be shown to him. The search bar is located on the header of every page and will display results of users in the database similar to the query from the user.

However, there is one thing i couldn't fix/make. If the user is logged out and tries to search for something in the search bar a login scheme will be displayed in the results page. However, after he successfully logs in he will not his/hers query will not be remembered and he will have to type again.